requirements.txt
streamlit>=1.25.0
google-genai>=1.1.0
python-dotenv>=1.0.0
textblob>=0.17.1
Pillow>=9.0.0
